Updated Italian translation
I added a few lines so that now every single word is correctly translated.
Please notice that "kb" in the string "Max cache file size (kb)" is "kB" I suppose. Am I right? I noticed that the capital letters of this string in most of .po files do not match the ones in prefs.js. Because of this, such sentence is not localized correctly. I fixed that in this italian .po file.
;)