Fix UTF8_STRING in registry.txt renders extension unresponsive on Gnome 46 (Wayland)
The idea of the fix comes from @gentoo-root, @AmeyaVS and @alephnull
comments in the original issue:

Changing "UTF8_STRING" to "text/plain;charset=utf-8" in registry.txt
seems to be fixing the problem after reloading the extension.

So let's use this.isText() instead of mimetype.startsWith('text/')) in
fromJSON text checking.

Without this patch "Text will be here" appears in the tray and extension
is unresponsive.

Fixes: #459

Signed-off-by: Pavel Tikhomirov <snorcht@gmail.com>